[Oztechan/TraceFit#72] Rename composeApp to client
