+++ client/src/androidMain/kotlin/com/oztechan/tracefit/client/Common.android.kt
@@ -1,1 +1,3 @@
+package com.oztechan.tracefit.client
+
 actual fun getPlatformName(): String = "Android"
