+++ client/src/androidMain/kotlin/com/oztechan/tracefit/client/MainActivity.kt
@@ -1,6 +1,5 @@
-package com.oztechan.tracefit
+package com.oztechan.tracefit.client
 
-import App
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
