+++ client/src/androidMain/res/mipmap-hdpi/ic_launcher.png
