+++ client/src/androidMain/res/mipmap-xxhdpi/ic_launcher.png
