+++ client/src/androidMain/res/mipmap-xxxhdpi/ic_launcher.png
