+++ client/src/commonMain/kotlin/com/oztechan/tracefit/client/App.kt
@@ -1,3 +1,5 @@
+package com.oztechan.tracefit.client
+
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
