+++ client/src/commonMain/kotlin/com/oztechan/tracefit/client/Common.kt
@@ -1,0 +1,3 @@
+package com.oztechan.tracefit.client
+
+expect fun getPlatformName(): String
