+++ client/src/commonMain/kotlin/com/oztechan/tracefit/client/Shared.kt
@@ -1,0 +1,3 @@
+package com.oztechan.tracefit.client
+
+const val GREETING = "Hello World!"
