+++ client/src/iosMain/kotlin/com/oztechan/tracefit/client/Common.ios.kt
@@ -1,1 +1,3 @@
+package com.oztechan.tracefit.client
+
 actual fun getPlatformName(): String = "iOS"
