+++ client/src/iosMain/kotlin/com/oztechan/tracefit/client/MainViewController.kt
@@ -1,3 +1,5 @@
+package com.oztechan.tracefit.client
+
 import androidx.compose.ui.window.ComposeUIViewController
 
 @Suppress("unused", "FunctionNaming") // used in iOS
