--- composeApp/src/commonMain/kotlin/Common.kt
@@ -1,1 +1,0 @@
-expect fun getPlatformName(): String
