--- composeApp/src/commonMain/kotlin/Shared.kt
@@ -1,1 +1,0 @@
-const val GREETING = "Hello World!"
