+++ iosApp/iosApp/ContentView.swift
@@ -1,6 +1,6 @@
 import UIKit
 import SwiftUI
-import ComposeApp
+import Client
 
 struct ComposeView: UIViewControllerRepresentable {
     func makeUIViewController(context: Context) -> UIViewController {
